[core] Add `plugins` sub-package
