+++ meshroom/core/plugins/manager.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 
-import glob
 import importlib
 import json
 import logging
 import os
-import re
 import sys
 
 from enum import Enum
@@ -15,8 +13,7 @@ from pathlib import Path
 from meshroom.common import BaseObject
 from meshroom.core import desc
 from meshroom.core.desc.attribute import ValueTypeErrors
-from meshroom import _MESHROOM_ROOT
-from meshroom.core.desc.node import _MESHROOM_COMPUTE_DEPS
+from meshroom.core.plugins.env import ProcessEnv
 
 
 def validateNodeDesc(nodeDesc: desc.BaseNode) -> list[tuple[str, ValueTypeErrors]]:
@@ -75,197 +72,6 @@ def formatNodeDescriptionErrorMessage(error: tuple[str, ValueTypeErrors]) -> str
     return f"Unknown error for parameter '{errMsg}'."
 
 
-class ProcessEnvType(Enum):
-    """ Supported process environments. """
-    DIRTREE = "dirtree",
-    REZ = "rez"
-
-
-class ProcessEnv(BaseObject):
-    """
-    Describes the environment required by a node's process.
-
-    Args:
-        folder: the source folder for the process.
-        configEnv: the dictionary containing the environment variables defined in a configuration file
-                   for the process to run.
-        pluginName: the name of the plugin object.
-        envType: (optional) the type of process environment.
-        uri: (optional) the Unique Resource Identifier to activate the environment.
-    """
-
-    def __init__(self, folder: str, configEnv: dict[str, str], pluginName: str,
-                 envType: ProcessEnvType = ProcessEnvType.DIRTREE, uri: str = ""):
-        super().__init__()
-        self._folder: str = folder
-        self._configEnv: dict[str: str] = configEnv
-        self.pluginName: str = pluginName
-        self._processEnvType: ProcessEnvType = envType
-        self.uri: str = uri
-        self._env: dict = None
-
-    def getEnvDict(self) -> dict:
-        """ Return the environment dictionary if it has been modified, None otherwise. """
-        return self._env
-
-    def getCommandPrefix(self) -> str:
-        """ Return the prefix to the command line that will be executed by the process. """
-        return ""
-
-    def getCommandSuffix(self) -> str:
-        """ Return the suffix to the command line that will be executed by the process. """
-        return ""
-
-
-class DirTreeProcessEnv(ProcessEnv):
-    """
-    """
-    def __init__(self, folder: str, configEnv: dict[str: str], pluginName: str):
-        super().__init__(folder, configEnv, pluginName, envType=ProcessEnvType.DIRTREE)
-
-        # If there is a virtual environment, it is expected to be named "venv".
-        # Beside the virtual environment, a standard "bin"/"lib"/"lib64" hierarchy at
-        # the top level of the plugin folder is expected.
-        venvFolder = Path(folder, "venv")
-
-        # Find all the libs that are not directly at the "lib*"-level
-        envLibPaths = glob.glob(f'{folder}/lib*/python[0-9].[0-9]*/site-packages',
-                                recursive=False)
-        venvLibPaths = glob.glob(f'{venvFolder}/lib*/python[0-9].[0-9]*/site-packages',
-                                 recursive=False)
-
-        self.binPaths: list = [str(Path(folder, "bin")), str(Path(venvFolder, "bin"))]
-        self.libPaths: list = [str(Path(folder, "lib")), str(Path(folder, "lib64")),
-                               str(Path(venvFolder, "lib")), str(Path(venvFolder, "lib64"))]
-        self.pythonPaths: list = [str(Path(folder)), str(Path(venvFolder))] + \
-                                 self.binPaths + envLibPaths + venvLibPaths
-
-        if sys.platform == "win32":
-            # For Windows platforms, try and include the content of the virtual env if it exists
-            # The virtual env is expected to be named "venv"
-            venvLibPath = Path(venvFolder, "Lib", "site-packages")
-            if venvLibPath.exists():
-                self.pythonPaths.append(venvLibPath.as_posix())
-        else:
-            # For Linux platforms, lib paths may need to be discovered recursively to be properly
-            # added to LD_LIBRARY_PATH
-            extraLibPaths = []
-            regex = re.compile(r"^lib(\d{2})?$")
-            for envPath in envLibPaths + venvLibPaths:
-                for path, directories, _ in os.walk(envPath):
-                    for directory in directories:
-                        if re.match(regex, directory):
-                            extraLibPaths.append(os.path.join(path, directory))
-            self.libPaths = self.libPaths + extraLibPaths
-
-        # Setup the environment dictionary
-        self._env = os.environ.copy()
-        self._env["PYTHONPATH"] = os.pathsep.join(
-            [f"{_MESHROOM_ROOT}"] + self.pythonPaths + [os.getenv('PYTHONPATH', '')])
-        self._env["LD_LIBRARY_PATH"] = f"{os.pathsep.join(self.libPaths)}{os.pathsep}{os.getenv('LD_LIBRARY_PATH', '')}"
-        self._env["PATH"] = f"{os.pathsep.join(self.binPaths)}{os.pathsep}{os.getenv('PATH', '')}"
-
-        for k, val in self._configEnv.items():
-            # Preserve user-defined environment variables:
-            # manually set environment variable values take precedence over config file defaults.
-            if k in self._env:
-                continue
-
-            self._env[k] = val
-
-
-class RezProcessEnv(ProcessEnv):
-    """
-    """
-
-    REZ_DELIMITER_PATTERN = re.compile(r"-|==|>=|>|<=|<")
-
-    def __init__(self, folder: str, configEnv: dict[str: str], pluginName: str, uri: str = ""):
-        if not uri:
-            raise RuntimeError("Missing name of the Rez environment needs to be provided.")
-        super().__init__(folder, configEnv, pluginName, envType=ProcessEnvType.REZ, uri=uri)
-
-    def resolveRezSubrequires(self) -> list[str]:
-        """
-        Return the list of packages defined for the node execution. These execution packages are
-        named subrequires.
-        Note: If a package does not have a version number, the version is aligned with the main
-        Meshroom environment (if this package is defined).
-        """
-        if os.getenv(f"{self.uri.upper()}_{self.pluginName.upper()}_SUBREQUIRES"):
-            subrequires = os.environ.get(f"{self.uri.upper()}_{self.pluginName.upper()}_SUBREQUIRES", "").split(os.pathsep)
-        else:
-            subrequires = os.environ.get(f"{self.uri.upper()}_SUBREQUIRES", "").split(os.pathsep)
-        if not subrequires:
-            return []
-
-        packages = []
-        # Packages that are resolved in the current environment
-        currentEnvPackages = []
-        resolvedVersions = {}
-        if "REZ_USED_RESOLVE" in os.environ:
-            resolvedPackages = os.getenv("REZ_USED_RESOLVE", "").split()
-            for package in resolvedPackages:
-                if package.startswith("~"):
-                    continue
-                currentEnvPackages.append(package)
-                name, version = self.REZ_DELIMITER_PATTERN.split(package, maxsplit=1)
-                resolvedVersions[name] = version
-        logging.debug("Packages in the current environment: " + ", ".join(currentEnvPackages))
-
-        # Take packages with the set versions for those which have one, and try to take packages
-        # in the current environment (if they are resolved in it)
-        for package in subrequires:
-            packageTuple = self.REZ_DELIMITER_PATTERN.split(package, maxsplit=1)
-            if len(packageTuple) == 1:
-                # Only the package name in the subrequires.
-                # Search for a corresponding version in the parent environment.
-                packageName = packageTuple[0]
-                parentResolvedVersion = resolvedVersions.get(packageName)
-                if parentResolvedVersion:
-                    packages.append(f"{packageName}=={parentResolvedVersion}")
-                else:
-                    packages.append(package)
-            elif len(packageTuple) == 2:
-                # The subrequires ask for a specific version
-                packages.append(package)
-
-        def extractPackageName(packageString: str) -> str:
-            return self.REZ_DELIMITER_PATTERN.split(packageString, maxsplit=1)[0]
-        packageNames = [extractPackageName(package) for package in packages]
-
-        for package in _MESHROOM_COMPUTE_DEPS:
-            # For packages that are required by meshroom_compute, do not specify any version
-            # or align it with Meshroom's: the version will be found during the resolution of
-            # the environment based on the other packages.
-            # If any of these packages is already part of the environment a plugin's dependency,
-            # do not add it
-            if package not in packageNames:
-                packages.append(package)
-
-        logging.debug("Packages for the execution environment: " + ", ".join(packages))
-        return packages
-
-    def getCommandPrefix(self):
-        # TODO: make Windows-compatible
-
-        # Use the PYTHONPATH from the subrequires' environment (which will only be resolved once
-        # inside the execution environment) and add MESHROOM_ROOT and the plugin's folder itself
-        # to it
-        pythonPaths = f"{os.pathsep.join(['$PYTHONPATH', f'{_MESHROOM_ROOT}', f'{self._folder}'])}"
-
-        return f"rez env {' '.join(self.resolveRezSubrequires())} -c 'PYTHONPATH={pythonPaths} "
-
-    def getCommandSuffix(self):
-        return "'"
-
-
-def processEnvFactory(folder: str, configEnv: dict[str: str], pluginName: str, envType: str = "dirtree", uri: str = "") -> ProcessEnv:
-    if envType == "dirtree":
-        return DirTreeProcessEnv(folder, configEnv, pluginName)
-    return RezProcessEnv(folder, configEnv, pluginName, uri=uri)
-
-
 class NodeProviderStatus(Enum):
     """
     Loading status for NodeProvider objects.
@@ -275,6 +81,151 @@ class NodeProviderStatus(Enum):
     DESC_ERROR = 2  # The node provider exists but has an invalid description
     LOADING_ERROR = 3  # The node provider exists and is valid but could not be successfully loaded
     ERROR = 4  # Error when importing the node provider from its module
+
+
+class NodeProvider(BaseObject):
+    """
+    Based on a node description, a NodeProvider represents a loadable node.
+
+    Members:
+        plugin: the Plugin object that contains this node provider
+        path: absolute path to the file containing the node's description
+        nodeDescriptor: the description of the node
+        status: the loading status on the node provider
+        errors: the list of errors (if there are any) when validating the description
+                of the node or attempting to load it
+        processEnv: the environment required for the node provider's process. It can either
+                    be specific to this node provider, or be common for all the node providers within
+                    the plugin
+        timestamp: the timestamp corresponding to the last time the node description's file has been
+                   modified
+    """
+
+    def __init__(self, nodeDesc: desc.BaseNode, plugin: Plugin = None):
+        super().__init__()
+        self.plugin: Plugin = plugin
+        self.path: str = Path(getfile(nodeDesc)).resolve().as_posix()
+        self.nodeDescriptor: desc.BaseNode = nodeDesc
+        self.nodeDescriptor.provider = self
+
+        self.status: NodeProviderStatus = NodeProviderStatus.NOT_LOADED
+        self.errors: list[tuple[str, ValueTypeErrors]] = validateNodeDesc(nodeDesc)
+
+        if self.errors:
+            self.status = NodeProviderStatus.DESC_ERROR
+
+        self._processEnv = None
+        self._timestamp = os.path.getmtime(self.path)
+
+    def reload(self) -> bool:
+        """
+        Reload the node provider and update its status accordingly. If the timestamp of the node provider's
+        path has not changed since the last time the plugin has been loaded, then nothing will happen.
+
+        Returns:
+            bool: True if the node provider has successfully been reloaded (i.e. there was no error, and
+                  some changes were made since its last loading), False otherwise.
+        """
+        timestamp = 0.0
+        try:
+            timestamp = os.path.getmtime(self.path)
+        except FileNotFoundError:
+            self.status = NodeProviderStatus.ERROR
+            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: The path at {self.path} was not "
+                          f"not found.")
+            return False
+
+        if self._timestamp == timestamp:
+            logging.info(f"[Reload] {self.nodeDescriptor.__name__}: Not reloading. The node description "
+                         f"at {self.path} has not been modified since the last load.")
+            return False
+
+        try:
+            updated = importlib.reload(sys.modules.get(self.nodeDescriptor.__module__))
+        except Exception as exc:
+            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: {exc} ({type(exc).__name__})")
+            self.status = NodeProviderStatus.DESC_ERROR
+            return False
+        descriptor = getattr(updated, self.nodeDescriptor.__name__)
+
+        if not descriptor:
+            self.status = NodeProviderStatus.ERROR
+            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: The node description at {self.path} "
+                          f"was not found.")
+            return False
+
+        self.errors = validateNodeDesc(descriptor)
+        if self.errors:
+            self.status = NodeProviderStatus.DESC_ERROR
+            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: The node description at {self.path} "
+                          f"has description errors.")
+            return False
+
+        self.nodeDescriptor = descriptor
+        self.nodeDescriptor.provider = self
+        self._timestamp = timestamp
+        self.status = NodeProviderStatus.NOT_LOADED
+        logging.info(f"[Reload] {self.nodeDescriptor.__name__}: Successful reloading.")
+        return True
+
+    @property
+    def plugin(self):
+        """
+        Return the Plugin object that contains this node provider.
+        If the node provider has not been assigned to a plugin yet, this value will
+        be set to None.
+        """
+        return self._plugin
+
+    @plugin.setter
+    def plugin(self, plugin: Plugin):
+        """ Assign this node provider to a containing Plugin object. """
+        self._plugin = plugin
+
+    @property
+    def isUserPlugin(self):
+        """ Return whether the node plugin belongs to a user plugin. """
+        if self.plugin:
+            return self.plugin.isUserPlugin
+        return False
+
+    @property
+    def processEnv(self):
+        """"
+        Return the process environment that is specific to the node provider if it has any.
+        Otherwise, the Plugin's is returned.
+        """
+        if self._processEnv:
+            return self._processEnv
+        if self.plugin:
+            return self.plugin.processEnv
+        return None
+
+    @property
+    def runtimeEnv(self) -> dict:
+        """ Return the environment dictionary for the runtime. """
+        return self.processEnv.getEnvDict()
+
+    @property
+    def commandPrefix(self) -> str:
+        """ Return the command prefix for the NodeProvider's execution. """
+        if not self.processEnv:
+            return ""
+        return self.processEnv.getCommandPrefix()
+
+    @property
+    def commandSuffix(self) -> str:
+        """ Return the command suffix for the NodeProvider's execution. """
+        if not self.processEnv:
+            return ""
+        return self.processEnv.getCommandSuffix()
+
+    @property
+    def configFullEnv(self) -> dict[str: str]:
+        """ Return the plugin's full environment dictionary. """
+        if not self.plugin:
+            return {}
+        return self.plugin.configFullEnv
 
 
 class Plugin(BaseObject):
@@ -476,150 +427,6 @@ class Plugin(BaseObject):
         return name in self._nodeProviders
 
 
-class NodeProvider(BaseObject):
-    """
-    Based on a node description, a NodeProvider represents a loadable node.
-
-    Members:
-        plugin: the Plugin object that contains this node provider
-        path: absolute path to the file containing the node's description
-        nodeDescriptor: the description of the node
-        status: the loading status on the node provider
-        errors: the list of errors (if there are any) when validating the description
-                of the node or attempting to load it
-        processEnv: the environment required for the node provider's process. It can either
-                    be specific to this node provider, or be common for all the node providers within
-                    the plugin
-        timestamp: the timestamp corresponding to the last time the node description's file has been
-                   modified
-    """
-
-    def __init__(self, nodeDesc: desc.BaseNode, plugin: Plugin = None):
-        super().__init__()
-        self.plugin: Plugin = plugin
-        self.path: str = Path(getfile(nodeDesc)).resolve().as_posix()
-        self.nodeDescriptor: desc.BaseNode = nodeDesc
-        self.nodeDescriptor.provider = self
-
-        self.status: NodeProviderStatus = NodeProviderStatus.NOT_LOADED
-        self.errors: list[tuple[str, ValueTypeErrors]] = validateNodeDesc(nodeDesc)
-
-        if self.errors:
-            self.status = NodeProviderStatus.DESC_ERROR
-
-        self._processEnv = None
-        self._timestamp = os.path.getmtime(self.path)
-
-    def reload(self) -> bool:
-        """
-        Reload the node provider and update its status accordingly. If the timestamp of the node provider's
-        path has not changed since the last time the plugin has been loaded, then nothing will happen.
-
-        Returns:
-            bool: True if the node provider has successfully been reloaded (i.e. there was no error, and
-                  some changes were made since its last loading), False otherwise.
-        """
-        timestamp = 0.0
-        try:
-            timestamp = os.path.getmtime(self.path)
-        except FileNotFoundError:
-            self.status = NodeProviderStatus.ERROR
-            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: The path at {self.path} was not "
-                          f"not found.")
-            return False
-
-        if self._timestamp == timestamp:
-            logging.info(f"[Reload] {self.nodeDescriptor.__name__}: Not reloading. The node description "
-                         f"at {self.path} has not been modified since the last load.")
-            return False
-
-        try:
-            updated = importlib.reload(sys.modules.get(self.nodeDescriptor.__module__))
-        except Exception as exc:
-            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: {exc} ({type(exc).__name__})")
-            self.status = NodeProviderStatus.DESC_ERROR
-            return False
-        descriptor = getattr(updated, self.nodeDescriptor.__name__)
-
-        if not descriptor:
-            self.status = NodeProviderStatus.ERROR
-            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: The node description at {self.path} "
-                          f"was not found.")
-            return False
-
-        self.errors = validateNodeDesc(descriptor)
-        if self.errors:
-            self.status = NodeProviderStatus.DESC_ERROR
-            logging.error(f"[Reload] {self.nodeDescriptor.__name__}: The node description at {self.path} "
-                          f"has description errors.")
-            return False
-
-        self.nodeDescriptor = descriptor
-        self.nodeDescriptor.provider = self
-        self._timestamp = timestamp
-        self.status = NodeProviderStatus.NOT_LOADED
-        logging.info(f"[Reload] {self.nodeDescriptor.__name__}: Successful reloading.")
-        return True
-
-    @property
-    def plugin(self):
-        """
-        Return the Plugin object that contains this node provider.
-        If the node provider has not been assigned to a plugin yet, this value will
-        be set to None.
-        """
-        return self._plugin
-
-    @plugin.setter
-    def plugin(self, plugin: Plugin):
-        """ Assign this node provider to a containing Plugin object. """
-        self._plugin = plugin
-
-    @property
-    def isUserPlugin(self):
-        """ Return whether the node plugin belongs to a user plugin. """
-        if self.plugin:
-            return self.plugin.isUserPlugin
-        return False
-
-    @property
-    def processEnv(self):
-        """"
-        Return the process environment that is specific to the node provider if it has any.
-        Otherwise, the Plugin's is returned.
-        """
-        if self._processEnv:
-            return self._processEnv
-        if self.plugin:
-            return self.plugin.processEnv
-        return None
-
-    @property
-    def runtimeEnv(self) -> dict:
-        """ Return the environment dictionary for the runtime. """
-        return self.processEnv.getEnvDict()
-
-    @property
-    def commandPrefix(self) -> str:
-        """ Return the command prefix for the NodeProvider's execution. """
-        if not self.processEnv:
-            return ""
-        return self.processEnv.getCommandPrefix()
-
-    @property
-    def commandSuffix(self) -> str:
-        """ Return the command suffix for the NodeProvider's execution. """
-        if not self.processEnv:
-            return ""
-        return self.processEnv.getCommandSuffix()
-
-    @property
-    def configFullEnv(self) -> dict[str: str]:
-        """ Return the plugin's full environment dictionary. """
-        if not self.plugin:
-            return {}
-        return self.plugin.configFullEnv
-
 class PluginManager(BaseObject):
     """
     Manager for all the loaded Plugin objects as well as the loaded NodeProvider objects.
@@ -634,31 +441,6 @@ class PluginManager(BaseObject):
 
         self._plugins: dict[str: Plugin] = {}  # loaded plugins
         self._nodeProviders: dict[str: NodeProvider] = {}  # loaded node providers
-
-    def isLoaded(self, name: str) -> bool:
-        """
-        Return whether the node provider has been loaded already.
-
-        Args:
-            name: the name of the node provider.
-        """
-        return name in self._nodeProviders
-
-    def belongsToPlugin(self, name: str) -> Plugin:
-        """
-        Check whether the node provider belongs to a loaded plugin, independently from
-        whether it has been loaded or not.
-
-        Args:
-            name: the name of the node provider that needs to be searched for across plugins.
-
-        Returns:
-            Plugin | None: the Plugin the node belongs to if it exists, None otherwise.
-        """
-        for plugin in self._plugins.values():
-            if plugin.containsNodeProvider(name):
-                return plugin
-        return None
 
     def getPlugins(self) -> dict[str: Plugin]:
         """
@@ -726,6 +508,31 @@ class PluginManager(BaseObject):
                 for node in plugin.nodes.values():
                     self.unloadNodeProvider(node)
             del self._plugins[plugin.uname]
+
+    def belongsToPlugin(self, name: str) -> Plugin:
+        """
+        Check whether the node provider belongs to a loaded plugin, independently from
+        whether it has been loaded or not.
+
+        Args:
+            name: the name of the node provider that needs to be searched for across plugins.
+
+        Returns:
+            Plugin | None: the Plugin the node belongs to if it exists, None otherwise.
+        """
+        for plugin in self._plugins.values():
+            if plugin.containsNodeProvider(name):
+                return plugin
+        return None
+
+    def isLoaded(self, name: str) -> bool:
+        """
+        Return whether the node provider has been loaded already.
+
+        Args:
+            name: the name of the node provider.
+        """
+        return name in self._nodeProviders
 
     def getLoadedNodeProviders(self) -> dict[str: NodeProvider]:
         """
