+++ tests/test_submit.py
@@ -13,8 +13,8 @@ from .utils import registerNodeDesc
 import meshroom
 from meshroom.core import pluginManager, loadClassesNodes, loadSubmitters, registerSubmitter, meshroomFolder
 from meshroom.core.graph import Graph
-from meshroom.core.plugins import Plugin
 from meshroom.core.node import Node, Status
+from meshroom.core.plugins.manager import Plugin
 from meshroom.core.submitter import jobManager
 from meshroom.core.submitter import OrderedTask, OrderedTasks, OrderedTaskType
 from meshroom.submitters.localFarm.localFarmSubmitter import LocalFarmSubmitter, LocalFarmJob
