+++ tests/utils.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 import meshroom
 from meshroom.core import desc, pluginManager, loadPluginFolder
-from meshroom.core.plugins import NodeProvider
+from meshroom.core.plugins.manager import NodeProvider
 
 import os
